requirements.txt
discord.py
aiohttp
asyncio
